# openapi.json
{"openapi":"3.1.0","info":{"title":"SLOPESHIELD NER API","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/status":{"get":{"summary":"Status","operationId":"status_api_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/predictions/simulate":{"post":{"tags":["predictions"],"summary":"Simulate","operationId":"simulate_api_predictions_simulate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimulationIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/predictions/status":{"get":{"tags":["predictions"],"summary":"Model Status","operationId":"model_status_api_predictions_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/locations":{"get":{"tags":["locations"],"summary":"Locations","operationId":"locations_api_locations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/LocationOut"},"type":"array","title":"Response Locations Api Locations Get"}}}}}}},"/api/locations/{location_id}":{"get":{"tags":["locations"],"summary":"Location","operationId":"location_api_locations__location_id__get","parameters":[{"name":"location_id","in":"path","required":true,"schema":{"type":"string","title":"Location Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sensors/status":{"get":{"tags":["sensors"],"summary":"Sensor Status","operationId":"sensor_status_api_sensors_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/alerts":{"get":{"tags":["alerts"],"summary":"Alerts","operationId":"alerts_api_alerts_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"active","title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/{alert_id}/resolve":{"patch":{"tags":["alerts"],"summary":"Resolve Alert","operationId":"resolve_alert_api_alerts__alert_id__resolve_patch","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"integer","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/summary":{"get":{"tags":["reports"],"summary":"Report Summary","operationId":"report_summary_api_reports_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports/csv":{"get":{"tags":["reports"],"summary":"Report Csv","operationId":"report_csv_api_reports_csv_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports/html":{"get":{"tags":["reports"],"summary":"Report Html","operationId":"report_html_api_reports_html_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports/pdf":{"get":{"tags":["reports"],"summary":"Report Pdf","operationId":"report_pdf_api_reports_pdf_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports/problems-pdf":{"get":{"tags":["reports"],"summary":"Problems Pdf","operationId":"problems_pdf_api_reports_problems_pdf_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/data/risk":{"get":{"tags":["data"],"summary":"Risk","operationId":"risk_api_data_risk_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/data/rainfall":{"get":{"tags":["data"],"summary":"Rainfall","operationId":"rainfall_api_data_rainfall_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/data/historical":{"get":{"tags":["data"],"summary":"Historical Points","operationId":"historical_points_api_data_historical_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/data/summary":{"get":{"tags":["data"],"summary":"Summary","operationId":"summary_api_data_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/data/status":{"get":{"tags":["data"],"summary":"Status","operationId":"status_api_data_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/data/weather/{location_id}":{"get":{"tags":["data"],"summary":"Weather","operationId":"weather_api_data_weather__location_id__get","parameters":[{"name":"location_id","in":"path","required":true,"schema":{"type":"string","title":"Location Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/data/weather/{location_id}/history":{"get":{"tags":["data"],"summary":"Weather History","operationId":"weather_history_api_data_weather__location_id__history_get","parameters":[{"name":"location_id","in":"path","required":true,"schema":{"type":"string","title":"Location Id"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":7,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/data/layers":{"get":{"tags":["data"],"summary":"Layers","operationId":"layers_api_data_layers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/monitoring/status":{"get":{"tags":["monitoring"],"summary":"Status","operationId":"status_api_monitoring_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/monitoring/cycle":{"post":{"tags":["monitoring"],"summary":"Monitoring Cycle","operationId":"monitoring_cycle_api_monitoring_cycle_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LocationOut":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"state":{"type":"string","title":"State"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"riskLevel":{"type":"string","title":"Risklevel"},"probability":{"type":"number","title":"Probability"},"rainfall":{"type":"number","title":"Rainfall"},"slope":{"type":"number","title":"Slope"},"elevation":{"type":"number","title":"Elevation"},"landCover":{"type":"string","title":"Landcover"},"geology":{"type":"string","title":"Geology"},"historicalLandslides":{"type":"integer","title":"Historicallandslides"},"trend":{"type":"string","title":"Trend"},"modelSource":{"type":"string","title":"Modelsource","default":"demo"}},"type":"object","required":["id","name","state","latitude","longitude","riskLevel","probability","rainfall","slope","elevation","landCover","geology","historicalLandslides","trend"],"title":"LocationOut"},"LoginIn":{"properties":{"name":{"type":"string","title":"Name","default":"User"},"email":{"type":"string","title":"Email"}},"type":"object","required":["email"],"title":"LoginIn"},"PredictionOut":{"properties":{"probability":{"type":"number","title":"Probability"},"riskLevel":{"type":"string","title":"Risklevel"},"modelSource":{"type":"string","title":"Modelsource"},"factors":{"items":{"type":"string"},"type":"array","title":"Factors","default":[]},"simulated":{"type":"boolean","title":"Simulated","default":false},"features":{"additionalProperties":{"type":"number"},"type":"object","title":"Features","default":{}}},"type":"object","required":["probability","riskLevel","modelSource"],"title":"PredictionOut"},"SimulationIn":{"properties":{"locationId":{"type":"string","title":"Locationid"},"rainfall":{"type":"number","maximum":1000.0,"minimum":0.0,"title":"Rainfall"}},"type":"object","required":["locationId","rainfall"],"title":"SimulationIn"},"UserOut":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"authenticated":{"type":"boolean","title":"Authenticated","default":true}},"type":"object","required":["name","email"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}